projects/03_coffee_shop_full_stack/starter_code/backend/requirements.txt
astroid==2.5.6
click==8.0.1
ecdsa==0.17.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.0
isort==5.8.0
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
pycryptodome==3.3.1
pylint==2.8.3
python-jose-cryptodome==1.3.2
six==1.16.0
SQLAlchemy==1.4.18
toml==0.10.2
typed-ast==1.4.3
Werkzeug==2.0.1
wrapt==1.12.1
